requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pillow>=10.0.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.30.0
pytest>=7.4.0
httpx>=0.24.0
